LocaleMap announce initialize and Trigger support improved.
LocaleMap now announces the initialize event for "pre-creationComplete" activities and properly dipatches a change event when the LocaleMap owner changes. 

The Injector classes now have better support for state changes by "scanning" the target hierarchy for any target or parent that has a valid <states/> array, that owner is the source of the STATE_CHANGE events... Injectors now add listeners with "smarts".