Improved default LoggingTarget and added support for ThunderBoltTarget.
Added support to use the ThunderBoltTarget to log to Firebug console. TraceTarget is still the default. But ThunderBoltTarget allows log viewing even when not debugging.
Fixed LocaleLogger with proper defaultFilters and added correct support for adding new filters.
Included the ThunderBoltAS3 source (with minor modifications) in my custom Swiz fork.